requirements.txt
httpx>=0.28.1
pillow>=11.0.0
python-dotenv>=1.2.1
requests>=2.32.5
langchain
langchain-openai
langchain-core
langgraph
langgraph-checkpoint-postgres
psycopg[binary,pool]
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
prometheus-fastapi-instrumentator>=7.0.0
prometheus-client>=0.20.0

